Changed the use of "firmware" to more conventional "software"
Great project! As someone who still is on WinForms (and loving it), I hope someday to see how you did the UI in WPF.

I changed the use of "firmware" to "software" to be more consistent.  Only a couple of tiny parts of Vector are called firmware -- the cube firmware and the body board firmware.